Fixed build issue with avrdude
avrdude-5.10 seems to fail building on Mac OS 10.6.8 (64-bit) with a
missing file config_gram.h. Oddly, that file does exist in obj-avr.
Updating to 5.11.1 fixed the issue.

Also, dropped adding the PREFIX to the PATH since PREFIX/bin is also
added to the PATH.

Dropped saving PWD to OLDPWD since that is a variable internal to bash
and is also useless to execute prior to the end of a script. This also
cleaned up an error that was being reported.