Add build.board property
This causes the ARDUINO_PINOCCIO macro to be defined when the Pinoccio
boards are selected. This does not distinguish between the 256RFR2 and
128RFR2 versions, but since the later was only a prototype, and the
regular CPU defines are also available, this shouldn't be a problem (the
alternative would be to check for both boards everywhere in the code,
which isn't nice either).