Use the regular arduino core
Previously, the pinoccio boards used an older copy of the regular
arduino core that is included in this repository. By explicitely listing
the core as "arduino:arduino" (instead of the previous implicit
pinoccio:arduino"), it uses the arduino core from the hardware/arduino
folder, i.e., the regular arduino core.

The variant folder is still taken ofr the hardware/pinoccio/variants
folder as before.