Define __PROG_TYPES_COMPAT__
The regular arduino core uses some types like prog_char, which are
deprecated on newer avr-libc versions. Since the Pinoccio requires a
newer toolchain than is included with Arduino, this breaks.

A fix is pending in the Arduino repo, but until it is in a released IDE
version, this workaround should fix things. See also
https://github.com/arduino/Arduino/pull/1695