Clean up pin comments in pins_arduino.h
There used to be three incomplete sets of comments in each of the three
arrays that are indexed by pin numbers. Now, there is a single comment
block near the top that lists all digital pins with their corresonding
functions (according to the datasheet) and pin label on the Pinoccio
Scout board.