Remove bootloader source and old versions
This firmware is now tracked in the hardware-pinoccio repository at
https://github.com/Pinoccio/hardware-pinoccio/tree/master/firmware/bootloader
(look in a tagged version to get a compiled version).

A single compiled version is still present, so "burn bootloader" in the
Arduino IDE still works. This version is versioned v1.0, referring to
the board version and tag in the hardware-pinoccio repository. It is
identical to what was previously called 0.51.